[PIK-32] Add DB config. Add connection between roles and users. Add security config
