--- src/main/java/edu/pw/eiti/pik/model/User.java
@@ -1,36 +1,0 @@
-package edu.pw.eiti.pik.model;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-import javax.persistence.*;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-@Entity
-@Data
-@EqualsAndHashCode
-@Inheritance(strategy = InheritanceType.JOINED)
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-
-    @NotBlank
-    @Email
-    private String email;
-
-    @NotBlank
-    private String password;
-
-    @NotBlank
-    private String name;
-
-    private String phoneNo;
-
-
-    public User() { }
-
-}
-
