+++ src/main/resources/application.properties
@@ -1,0 +1,12 @@
+# H2
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
+spring.h2.console.settings.trace=false
+spring.h2.console.settings.web-allow-others=true
+
+# Datasource
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.datasource.url=jdbc:h2:file:~/.pik/test/data
+spring.datasource.username=sa
+spring.datasource.password=sa
+spring.datasource.driver-class-name=org.h2.Driver
